requirements.txt
Flask
Flask-Caching
Flask-DebugToolbar
Flask-Login
Flask-SQLAlchemy
sqlalchemy_utils
pytest
python-dotenv